LayoutTests/media/audio-constructor.html

<body>
<p>Test that Audio() object loads the resource after src attribute is set and load() is called.</p>

<script src=video-test.js></script>

<script>
    var audio = new Audio();
    media = audio;

    testExpected("audio instanceof HTMLAudioElement", true);

    waitForEvent("loadstart", function () { testExpected("relativeURL(audio.currentSrc)", "content/test.wav"); });

    waitForEventAndEnd("load");

    audio.src = "content/test.wav";
    run("audio.load()");
</script>
